POSTMORTEM on Attempt to quash repeat children within ancestors.
i.e. "in (in)" occurs within "inin (inin)" twice, so the former should in theory be decremented twice per instance of the latter. In practice, this results in a cat-and-mouse game of adjustments for successively overcounting and undercounting. For example, matching input "asinine" with the substrings of "piccinini" yields an overlap on "piccin[in]i" due to "inin" and "in." The real proper way to fix this is to expand the optimized dict to include subsuming entries and their match indices, but it would be a ton of refactoring for something that likely has no meaningful impact on output quality.